requirement.txt

Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-allauth==65.8.1
requests==2.31.0
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.14.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt
